app/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
Booktype==1.5
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
Django==5.1.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.9
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
redis==5.2.1
simplejson==3.20.1
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
